Moved threadsafe header into baxterTictacToe library. Changed everything else accordingly
